pyproject.toml
[tool.poetry]
name = "ing-transactions"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "5.1.2"
djangorestframework = "3.15.2"
django-cors-headers = "^4.6.0"
mozilla-django-oidc = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
